Rearrange travis.yml steps
This change is meant to make the travis-ci code more readable and look 
less like spaghetti. 
I arranged all the shared steps in the order in which they actually occur. 
I reduced the amount of repeated code to what I think is the bare minimum.

Although unnecessary, I hope this commit is an improvement.