Improve regex pattern that checks python version
The travis-ci script uses grep to determine which version of Python
is used to run Conan, and it intentionally fails to build if it is configured
to use the wrong version. 

The last commit of this file would have allowed any Python with a '3' 
in the version, for instance Python 2.7.3, to evade this check. 

This commit resolves that problem.