Fix Input.hpp build failure on Visual Studio
Input.hpp is missing the 'optional' header; this change adds it.

Visual Studio emits a warning about casting an enum to an unsigned integer
on line 334; this change silences the warning with a static cast.

Visual Studio emits a warning about an unused variable `index` on line 403;
this variable is clearly used on the next line, but VS doesn't recognize it.
This change silences the warning with a [[maybe_unused]] annotation.
This is not an optimal solution to the problem, but it's better than disabling
"warnings as errors", and it results in a successful build.